[refactor] 캐러셀 버튼 생성 함수 모듈화
- utils.js에 모듈화 함수 작성
