+++ index.html
@@ -1,4 +1,4 @@
-<!doctype html>
+<!DOCTYPE html>
 <html lang="kr">
   <head>
     <meta charset="UTF-8" />
@@ -199,6 +199,7 @@
         </ul>
         <button class="carousel-arrow left playlist__arrow--prev" aria-label="이전"></button>
       </section>
+
       <section>
         <div class="header-area">
           <h2>📍 날씨에 어울리는 장소를 추천드릴게요</h2>
@@ -206,12 +207,8 @@
         <div class="carousel-area">
           <div class="carousel-outer" id="carouselOuter">
             <ul class="place-list" id="carousel"></ul>
-            <button class="carousel-arrow left" id="arrowLeft" aria-label="이전">
-              <span class="arrow-svg"></span>
-            </button>
-            <button class="carousel-arrow right" id="arrowRight" aria-label="다음">
-              <span class="arrow-svg"></span>
-            </button>
+            <button class="carousel-arrow left" id="arrowLeft" aria-label="이전"></button>
+            <button class="carousel-arrow right" id="arrowRight" aria-label="다음"></button>
           </div>
         </div>
       </section>
