+++ src/components/placeRecommand/woohyeun.css
@@ -181,17 +181,6 @@ body {
   box-shadow: 0 0.25rem 1rem rgba(0, 0, 0, 0.1);
 }
 
-.arrow-svg {
-  width: 100%;
-  height: 100%;
-  display: inline-block;
-
-  & svg {
-    width: 100%;
-    height: 100%;
-  }
-}
-
 @media (max-width: 640px) {
   .header-area h2 {
     text-align: center;
@@ -206,10 +195,7 @@ body {
     transform: translateY(-50%) scale(0.93);
     box-shadow: 0 0.25rem 1rem rgba(0, 0, 0, 0.1);
   }
-  .arrow-svg {
-    width: 1.25rem;
-    height: 1.25rem;
-  }
+
   .carousel-outer {
     min-width: 0;
   }
