+++ src/js/utils.js
@@ -44,3 +44,40 @@ export function updateHighlight(items, currentIndex) {
     }
   });
 }
+
+// 버튼에 이미지 삽입 함수
+export function updateArrows() {
+  let size;
+
+  // 브라위저 너비에 따라 size 변경
+  if (window.innerWidth <= 1000) size = 38;
+  if (window.innerWidth <= 640) size = 28;
+
+  // .carousel-arrow 버튼 노드리스트를 불러옴
+  const buttons = document.querySelectorAll(".carousel-arrow");
+  // 버튼 노드리스트가 없으면 빠른 반환
+  if (buttons.length === 0) return;
+
+  // forEach문을 사용하여 버튼 이미지 삽입
+  buttons.forEach((button) => {
+    if (button.classList.contains("left")) {
+      button.innerHTML = getArrowSvg("left", size);
+    } else if (button.classList.contains("right")) {
+      button.innerHTML = getArrowSvg("right", size);
+    }
+  });
+}
+
+// 화살표 SVG 이미지 생성 함수
+function getArrowSvg(direction, size = 60) {
+  const arrow =
+    direction === "left"
+      ? `<polyline points="50,25 30,40.5 50,56" fill="none" stroke="black" stroke-width="6" stroke-linecap="round" stroke-linejoin="round"/>`
+      : `<polyline points="30,25 50,40.5 30,56" fill="none" stroke="black" stroke-width="6" stroke-linecap="round" stroke-linejoin="round"/>`;
+  return `
+    <svg width="${size}" height="${size}" viewBox="0 0 80 81" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <circle cx="40" cy="40.5" r="40" fill="#ffffff"/>
+      ${arrow}
+    </svg>
+  `;
+}
